# client/public/swagger.json
{
  "openapi": "3.0.0",
  "info": {
    "description": "david energy",
    "version": "1.0.0",
    "title": "dave-energy"
  },
  "servers": [
    {
      "url": "https://david-energy-staging.herokuapp.com"
    },
    {
      "url": "http://127.0.0.1:5000"
    }
  ],
...
  "paths": {
...
    "/building/{building_id}/energy-star-rating": {
      "get": {
        "tags": [
          "Building"
        ],
        "summary": "Get building energy star rating. NOTE: Log-in required",
        "description": "Returns all the buildings energy star rating associated with the building",
        "parameters": [
          {
            "name": "building_id",
            "in": "path",
            "description": "pass building_id",
            "schema": {
              "type": "integer",
              "example": 5
            }
          },
          {
            "name": "year",
            "in": "path",
            "description": "pass year (int, optional): year=2019",
            "schema": {
              "type": "integer",
              "example": 2019
            }
          }
        ],
        "responses": {
          "200": {
            "description": "successful operation",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/building-energy-star-rating"
                  }
                }
              }
            }
          }
        }
      }
    },
    "/building/{building_id}/energy-demand": {
      "get": {
        "tags": [
          "Building"
        ],
        "summary": "Get building energy demand",
        "description": "Devices information for a particular building.",
        "parameters": [
          {
            "name": "building_id",
            "in": "path",
            "description": "pass building_id",
            "schema": {
              "type": "integer",
              "example": 5
            }
          },
          {
            "name": "energy_type",
            "in": "query",
            "description": "pass energy type like electric, steam, and gas etc.",
            "schema": {
              "type": "string",
              "example": "electric"
            }
          },
          {
            "name": "from_time",
            "in": "query",
            "description": "pass from_time",
            "schema": {
              "type": "string",
              "example": "2020-04-24T00:00:00.000000"
            }
          },
          {
            "name": "to_time",
            "in": "query",
            "description": "pass to_time",
            "schema": {
              "type": "string",
              "example": "2020-05-11T00:00:00.000000"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "successful operation",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/energy-demand"
                  }
                }
              }
            }
          }
        }
      }
    },
...
    "/building/{building_id}/event-scheduler": {
      "get": {
        "tags": [
          "Building"
        ],
        "summary": "Get building event",
        "description": "dict: Events for a particular building.",
        "parameters": [
          {
            "name": "building_id",
            "in": "path",
            "description": "pass building_id",
            "schema": {
              "type": "integer",
              "example": 5
            }
          },
          {
            "name": "from_time",
            "in": "query",
            "description": "pass from_time (datetime, optional): from_time=2020-04-24T00:00:00.000000 ",
            "schema": {
              "type": "string",
              "example": "2020-04-24T00:00:00.000000"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "successful operation",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/dispatch"
                  }
                }
              }
            }
          }
        }
      }
    },
...
    "/building/{building_id}/billing-cycle": {
      "get": {
        "tags": [
          "Building"
        ],
        "summary": "Get building billing-cycle",
        "description": "",
        "parameters": [
          {
            "name": "building_id",
            "in": "path",
            "description": "pass building_id",
            "schema": {
              "type": "integer",
              "example": 5
            }
          },
          {
            "name": "from_time",
            "in": "query",
            "description": "pass from_time (datetime, optional): from_time=2020-04-24T00:00:00.000000 ",
            "schema": {
              "type": "string",
              "example": "2020-04-24T00:00:00.000000"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "successful operation",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                    "$ref": "#/components/schemas/billing-cycle"
                }
              }
            }
          }
        }
      }
    },
    "/building/{building_id}/billing-information": {
      "get": {
        "tags": [
          "Building"
        ],
        "summary": "Get building billing-information",
        "description": "",
        "parameters": [
          {
            "name": "building_id",
            "in": "path",
            "description": "pass building_id",
            "schema": {
              "type": "integer",
              "example": 5
            }
          },
          {
            "name": "from_time",
            "in": "query",
            "description": "pass from_time (datetime, optional): from_time=2020-04-24T00:00:00.000000 ",
            "schema": {
              "type": "string",
              "example": "2020-04-24T00:00:00.000000"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "successful operation",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                    "$ref": "#/components/schemas/billing-information"
                }
              }
            }
          }
        }
      }
    },
...
    "/building/{building_id}/greenhouse-gas-emissions": {
      "get": {
        "tags": [
          "Building"
        ],
        "summary": "Get building greenhouse-gas-emissions",
        "description": "",
        "parameters": [
          {
            "name": "building_id",
            "in": "path",
            "description": "pass building_id",
            "schema": {
              "type": "integer",
              "example": 5
            }
          },
          {
            "name": "year",
            "in": "query",
            "description": "year (int, optional): year=2020 ",
            "schema": {
              "type": "integer",
              "example": "2020"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "successful operation",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/greenhouse-gas-emissions"
                  }
                }
              }
            }
          }
        }
      }
    },
...
    "/login": {
      "post": {
        "tags": [
          "User"
        ],
        "summary": "Log-in user",
        "description": "User can login via username and password",
        "requestBody": {
          "required": true,
          "content": {
            "application/json":{
              "schema": {
                "$ref": "#/components/schemas/login_request"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "successful operation",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "$ref": "#/components/schemas/user"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "text/plain": {
                "schema": {
                  "type": "string"
                }
              }
            }
          },
          "403": {
            "description": "Authentication Error",
            "content": {
              "text/plain": {
                "schema": {
                  "type": "string"
                }
              }
            }
          },
          "404": {
            "description": "Resource Not found",
            "content": {
              "text/plain": {
                "schema": {
                  "type": "string"
                }
              }
            }
          }
        }
      }
    },
    "/register": {
      "post": {
        "tags": [
          "User"
        ],
        "summary": "register user",
        "description": "User can register by passing username and password",
        "requestBody": {
          "required": true,
          "content": {
            "application/json":{
              "schema": {
                "$ref": "#/components/schemas/register"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "successful operation",
            "content": {
              "text/plain": {
                "schema": {
                  "type": "string",
                  "example": "user registered successfully"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "text/plain": {
                "schema": {
                  "type": "string"
                }
              }
            }
          },
          "409": {
            "description": "Resource Conflict Error",
            "content": {
              "text/plain": {
                "schema": {
                  "type": "string"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "building-energy-star-rating": {
        "type": "object",
        "properties": {
          "change": {
            "type": "integer",
            "example": 0
          },
          "month": {
            "type": "integer",
            "example": 1
          },
          "month_display": {
            "type": "string",
            "example": "J"
          },
          "value": {
            "type": "integer",
            "example": 65
          }
        }
      },
      "login_request": {
        "type": "object",
        "properties": {
        "email": {
            "type": "string",
            "example": "someone@gmail.com"
          },
          "password": {
            "type": "string",
            "example": "password"
          }
        }
      },
      "register": {
        "type": "object",
        "properties": {
        "email": {
            "type": "string",
            "example": "someone@gmail.com"
          },
          "password": {
            "type": "string",
            "example": "password"
          }
        }
      },
      "user": {
        "type": "object",
        "properties": {
        "buildings": {
            "type": "array",
            "example": "[]"
          },
          "company": {
            "type": "integer",
            "example": 5
          },
          "email": {
            "type": "string",
            "example": "rishabh@gmail.com"
          },
          "first_name": {
            "type": "string",
            "example": ""
          },
          "last_name": {
            "type": "string",
            "example": ""
          },
          "id": {
            "type": "integer",
            "example": "49"
          },
          "is_active": {
            "type": "boolean",
            "example": true
          },
          "user_role": {
            "type": "integer",
            "example": 1
          }
        }
      },
...
      "energy-demand": {
        "type": "object",
        "properties": {
        "electricity": {
            "type": "array",
            "items":{
              "type": "object",
              "properties": {
                "history": {
                  "type": "array",
                  "items": {
                    "type": "object",
                    "properties": {
                      "ts": {
                        "type": "string",
                        "example": "2020-08-24T00:00:00.000000"
                      },
                      "quantity": {
                        "type": "number",
                        "example": 56
                      }
                    }
                  }
                }
              }
            }
          },
          "target": {
            "type": "number",
            "example": 800.0
          }
        }
      },
...
      "billing-cycle": {
          "type": "object",
          "properties": {
            "days_left": {
              "type": "integer",
              "example":  -24
            },
            "from_date": {
              "type": "string",
              "example": "OCT 1"
            },
            "to_date": {
              "type": "string",
              "example": "OCT 31"
            }
          }
      },
      "billing-information": {
          "type": "object",
          "properties": {
            "billing-information": {
              "type": "object",
              "properties": {
                "days_left": {
                  "type": "integer",
                  "example": -24
                },
                "from_date": {
                  "type": "string",
                  "example": "OCT 1"
                },
                "to_date": {
                  "type": "string",
                  "example": "OCT 31"
                }
              }
            },
            "billing_day_total": {
              "type": "integer",
              "example": 773
            },
            "billing_peak": {
              "type": "object",
              "properties": {
                "quantity": {
                  "type": "integer",
                  "example": 221
                },
                "target": {
                  "type": "integer",
                  "example": 400
                },
                "targeted": {
                  "type": "integer",
                  "example": -45
                },
                "ts": {
                  "type": "string",
                  "example": "2020-10-02T09:45:00.000000"
                }
              }
            },
            "billing_total": {
              "type": "integer",
              "example": 9323
            }
          }
      },
...
      "greenhouse-gas-emissions": {
          "type": "object",
          "properties": {
            "month_display": {
              "type": "string",
              "example": "J"
            },
            "month_number": {
              "type": "integer",
              "example": 1
            },
            "quantity": {
              "type": "number",
              "example": 192624.0
            },
            "percentage": {
              "type": "number",
              "example": 0
            },
            "unit": {
              "type": "string",
              "example": "tCO2e"
            }
          }
      },
...
      "dispatch": {
          "type": "object",
          "properties": {
            "end_date": {
              "type": "string",
              "example":  "2020-09-11T16:00:02.420000"
            },
            "event_type": {
              "type": "string",
              "example": null
            },
            "mode": {
              "type": "string",
              "example": null
            },
            "point_id": {
              "type": "integer",
              "example": 415
            },
            "power_kw": {
              "type": "string",
              "example": null
            },
            "schedule_date": {
              "type": "string",
              "example": "2020-09-11T15:45:02.420000"
            }
          }
      },
...
    }
  }
}